Develop

+++ app/Http/Controllers/V1/RestaurantController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers\V1;
 
-use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
+use App\Models\User;
 use App\Models\Restaurant;
+use Illuminate\Http\Request;
+use App\Models\RestaurantPicture;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 
 class RestaurantController extends Controller
 {
@@ -24,7 +28,7 @@ class RestaurantController extends Controller
             'sort_order'    => 'nullable|in:asc,desc',
         ]);
 
-        $query = Restaurant::query()->with(['users', 'cousines', 'orders']);
+        $query = Restaurant::query()->with('users');
 
         if ($request->search) {
             $query = $query->where('name', 'like', "%$request->search%");
@@ -61,23 +65,51 @@ class RestaurantController extends Controller
     public function create(Request $request)
     {
         $this->validate($request, [
-            'user_id.*'           => 'required|integer|exists:users,id',
-            'cousine_type_id'     => 'required|integer|exists:cousine_types,id',
-            'name'                => 'required|alpha|max:20',
-            'address1'            => 'required|string|max:50',
-            'address2'            => 'nullable|string|max:50',
-            'zip_code'            => 'required|integer|min:6',
-            'phone'               => 'required|integer|min:10',
-            'profile_picture'     => 'required|mimes:jpg,jpeg,png,bmp,tiff'
+            'user_id'                       => 'required|integer|exists:users,id',
+            'cousine_type_id.*'             => 'required|integer|exists:cousine_types,id',
+            'name'                          => 'required|string|max:20',
+            'address1'                      => 'required|string|max:50',
+            'address2'                      => 'nullable|string|max:50',
+            'zip_code'                      => 'required|integer|min:6',
+            'phone'                         => 'required|integer|min:10',
+            'logo'                          => 'required|mimes:jpg,jpeg,png,bmp,tiff',
+            'type'                          => 'required|in:M,O',
+            'picture.*'                     => 'required|mimes:jpg,jpeg,png,bmp,tiff',
+            'stocks.*'                      => 'required|array',
+            'stocks.*.stock_type_id'        => 'required|integer|exists:stock_types,id',
+            'stocks.*.name'                 => 'required|alpha|max:20',
+            'stocks.*.available_quantity'   => 'required|numeric',
+            'stocks.*.minimum_quantity'     => 'required|numeric',
+
         ]);
 
-        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->profile_picture->getClientOriginalExtension();
-        $request->profile_picture->storeAs("public/pictures", $imageName);
+        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->logo->getClientOriginalExtension();
+        $request->logo->storeAs("public/pictures", $imageName);
 
-        $restaurant = Restaurant::create($request->only('cousine_type_id', 'name', 'address1', 'address2', 'phone', 'zip_code') + ['profile_picture' => $imageName]);
-        $restaurant->users()->syncWithoutDetaching($request->user_id);
+        $user = User::findOrFail($request->user_id);
 
-        return ok('Restaurant created successfully!',  $restaurant->load('users'));
+        $image = array();
+        if ($request->hasFile('picture')) {
+            foreach ($request->picture as $file) {
+                $image_name  =  str_replace(".", "", (string)microtime(true)) . '.' . $file->getClientOriginalExtension();
+                $upload_path =  'public/pictures';
+                $file->storeAs($upload_path, $image_name);
+                $image[] = [
+                    'restaurant_id' => $request->id,
+                    'picture'       => $image_name,
+                    'type'          => $request->type
+                ];
+            }
+        }
+        if ($user->role->name == "Owner") {
+            $restaurant = Restaurant::create($request->only('name', 'address1', 'address2', 'phone', 'zip_code') + ['logo' => $imageName]);
+            $restaurant->users()->syncWithoutDetaching([$request->user_id => ['is_owner' => true]]);
+            $restaurant->cousines()->syncWithoutDetaching($request->cousine_type_id);
+            $restaurant->pictures()->createMany($image);
+            $restaurant->stocks()->createMany($request->stocks);
+            return ok('Restaurant created successfully!',  $restaurant->load('users', 'pictures', 'cousines', 'stocks'));
+        }
+        return 'Restaurant can not created!';
     }
 
     /**
@@ -91,24 +123,47 @@ class RestaurantController extends Controller
     {
         $this->validate($request, [
             'user_id.*'           => 'nullable|integer|exists:users,id',
-            'cousine_type_id'     => 'nullable|integer|exists:cousine_types,id',
-            'name'                => 'required|alpha|max:20',
+            'cousine_type_id.*'   => 'nullable|integer|exists:cousine_types,id',
+            'name'                => 'required|string|max:20',
             'address1'            => 'nullable|string|max:50',
             'address2'            => 'nullable|string|max:50',
             'zip_code'            => 'nullable|integer|min:6',
             'phone'               => 'nullable|integer|min:10',
-            'profile_picture'     => 'required|mimes:jpg,jpeg,png,bmp,tiff'
+            'logo'                => 'required|mimes:jpg,jpeg,png,bmp,tiff',
+            'type'                => 'nullable|in:M,O',
+            'picture.*'           => 'nullable|mimes:jpg,jpeg,png,bmp,tiff'
         ]);
 
         $restaurant = Restaurant::findOrFail($id);
+        if ($restaurant->logo) {
+            Storage::delete("public/pictures/" . $restaurant->logo);
+            $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->logo->getClientOriginalExtension();
+            $request->logo->storeAs("public/pictures", $imageName);
+        }
+        $restaurant->update($request->only('name', 'address1', 'address2', 'phone', 'zip_code') + ['logo' => $imageName]);
+        $restaurant->cousines()->sync($request->cousine_type_id);
 
-        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->profile_picture->getClientOriginalExtension();
-        $request->profile_picture->storeAs("public/pictures", $imageName);
+        $picture = RestaurantPicture::findOrFail($id);
+        if ($picture->picture) {
+            $image = array();
+            if ($request->hasFile('picture')) {
+                Storage::delete("public/pictures/" . $picture->picture);
+                $restaurant->pictures()->delete();
+                foreach ($request->picture as $file) {
+                    $image_name =  str_replace(".", "", (string)microtime(true)) . '.' . $file->getClientOriginalExtension();
+                    $upload_path =  'public/pictures';
+                    $file->storeAs($upload_path, $image_name);
+                    $image[] = [
+                        'restaurant_id' => $request->id,
+                        'picture' => $image_name,
+                        'type' => $request->type
+                    ];
+                }
+            }
+        }
 
-        $restaurant->update($request->only('cousine_type_id', 'name', 'address1', 'address2', 'phone', 'zip_code') + ['profile_picture' => $imageName]);
-        $restaurant->users()->sync($request->user_id);
-
-        return ok('Restaurant updated successfully!',  $restaurant->load('users'));
+        $restaurant->pictures()->createMany($image);
+        return ok('Restaurant updated successfully!', $restaurant->load('cousines', 'pictures'));
     }
 
     /**
@@ -119,7 +174,7 @@ class RestaurantController extends Controller
      */
     public function get($id)
     {
-        $restaurant = Restaurant::with(['users', 'cousines', 'orders'])->findOrFail($id);
+        $restaurant = Restaurant::with(['users', 'cousines', 'pictures', 'orders'])->findOrFail($id);
 
         return ok('Restaurant retrieved successfully',  $restaurant);
     }
@@ -133,8 +188,11 @@ class RestaurantController extends Controller
     public function delete($id)
     {
         $restaurant = Restaurant::findOrFail($id);
-        if ($restaurant->users()->count() > 0) {
+        if ($restaurant->users()->count() > 0 && $restaurant->pictures()->count() > 0 && $restaurant->cousines()->count() > 0) {
             $restaurant->users()->detach();
+            $restaurant->cousines()->detach();
+            $restaurant->pictures()->delete();
+            $restaurant->stocks()->delete();
         }
         $restaurant->delete();
         return ok('Restaurant deleted successfully');
